requirements.txt
robotframework==3.0
robotframework-Selenium2Library
pyyaml